requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.1
python-ulid>=1.1.0
